House in Azerbaijan
Buy a new house/villa 500 sq/m in Baku Shuvelan settlement. The house is located near the road. 12 acres of land. There are 7 rooms in the 3-storey villa. On the 1st floor there are 2 bedrooms, a living room, a kitchen and 3 bathrooms. On the 2nd floor, there are 4 bedrooms, 4 bathrooms. 3rd floor attic. The rooms of the house have parquet floors. The house has a 3 meter high basement. The facade of the house is covered with aglay. Security cameras and an alarm are connected to the perimeter. There is a 11-meter-long, 4-meter-wide swimming pool in the yard with a capacity of 70 tons. Additionally, there is a garage for 4 cars, a gazebo, and a barbecue. Water, gas, electricity constant. Documents are in order.